+++ server/Library/ipc.js
@@ -1,27 +1,43 @@
-const log = require('../lib/logger')('Media:ipc')
-const Media = require('./Media')
+const Library = require('./Library')
+const log = require('../lib/logger')('Library:ipc')
+const throttle = require('@jcoreio/async-throttle')
 const {
-  MEDIA_ADD,
-  MEDIA_CLEANUP,
-  MEDIA_REMOVE,
-  MEDIA_UPDATE,
+  SCANNER_WORKER_STATUS,
+  LIBRARY_PUSH,
+  LIBRARY_MATCH,
   _SUCCESS,
   _ERROR,
 } = require('../../shared/actionTypes')
 
 const ACTION_HANDLERS = {
-  [MEDIA_ADD]: async ({ payload }) => Media.add(payload),
-  [MEDIA_CLEANUP]: Media.cleanup,
-  [MEDIA_REMOVE]: async ({ payload }) => Media.remove(payload),
-  [MEDIA_UPDATE]: async ({ payload }) => Media.update(payload),
+  [LIBRARY_MATCH]: async ({ payload }) => Library.match(payload),
 }
 
 /**
  * IPC action handler for scanner (db writes occur in server process)
  */
 module.exports = function (io) {
+  const emit = throttle(action => io.emit('action', action), 1000)
+
   process.on('message', async function (action) {
     const { type } = action
+
+    // broadcast scanner status updates
+    if (type === SCANNER_WORKER_STATUS) {
+      emit(action)
+
+      // scanner finished?
+      if (action.payload.isScanning === false) {
+        Library.setLibraryVersion()
+
+        io.emit('action', {
+          type: LIBRARY_PUSH,
+          payload: await Library.get(),
+        })
+      }
+
+      return
+    }
 
     if (typeof ACTION_HANDLERS[type] !== 'function') {
       log.debug('ignoring action: %s', type)