+++ server/Queue/Queue.js
@@ -1,5 +1,5 @@
 const path = require('path')
-const db = require('sqlite')
+const db = require('../lib/Database').db
 const sql = require('sqlate')
 
 class Queue {